move space view screenshot to its own module, kill dead code
