+++ crates/re_viewport/src/viewport.rs
@@ -10,7 +10,6 @@ use once_cell::sync::Lazy;
 
 use re_entity_db::EntityPropertyMap;
 use re_renderer::ScreenshotProcessor;
-use re_space_view::ScreenshotMode;
 use re_space_view::SpaceViewBlueprint;
 use re_ui::{Icon, ReUi};
 use re_viewer_context::{
@@ -19,6 +18,7 @@ use re_viewer_context::{
 };
 
 use crate::container::blueprint_id_to_tile_id;
+use crate::screenshot::handle_pending_space_view_screenshots;
 use crate::{
     add_space_view_or_container_modal::AddSpaceViewOrContainerModal, container::Contents,
     icon_for_container_kind, space_view_entity_picker::SpaceViewEntityPicker,
@@ -314,7 +314,9 @@ impl<'a, 'b> Viewport<'a, 'b> {
             while ScreenshotProcessor::next_readback_result(
                 ctx.render_ctx,
                 space_view.id.gpu_readback_id(),
-                |data, extent, mode| handle_pending_screenshots(space_view, data, extent, mode),
+                |data, extent, mode| {
+                    handle_pending_space_view_screenshots(space_view, data, extent, mode)
+                },
             )
             .is_some()
             {}
@@ -962,57 +964,5 @@ impl TabWidget {
             self.galley,
             label_color,
         );
-    }
-}
-
-// ----------------------------------------------------------------------------
-
-fn handle_pending_screenshots(
-    space_view: &SpaceViewBlueprint,
-    data: &[u8],
-    extent: glam::UVec2,
-    mode: ScreenshotMode,
-) {
-    // Set to clipboard.
-    #[cfg(not(target_arch = "wasm32"))]
-    re_viewer_context::Clipboard::with(|clipboard| {
-        clipboard.set_image([extent.x as _, extent.y as _], data);
-    });
-    if mode == ScreenshotMode::CopyToClipboard {
-        return;
-    }
-
-    // Get next available file name.
-    fn is_safe_filename_char(c: char) -> bool {
-        c.is_alphanumeric() || matches!(c, ' ' | '-' | '_')
-    }
-    let safe_display_name = space_view
-        .display_name_or_default()
-        .as_ref()
-        .replace(|c: char| !is_safe_filename_char(c), "");
-    let mut i = 1;
-    let filename = loop {
-        let filename = format!("Screenshot {safe_display_name} - {i}.png");
-        if !std::path::Path::new(&filename).exists() {
-            break filename;
-        }
-        i += 1;
-    };
-    let filename = std::path::Path::new(&filename);
-
-    match image::save_buffer(filename, data, extent.x, extent.y, image::ColorType::Rgba8) {
-        Ok(_) => {
-            re_log::info!(
-                "Saved screenshot to {:?}.",
-                filename.canonicalize().unwrap_or(filename.to_path_buf())
-            );
-        }
-        Err(err) => {
-            re_log::error!(
-                "Failed to safe screenshot to {:?}: {}",
-                filename.canonicalize().unwrap_or(filename.to_path_buf()),
-                err
-            );
-        }
     }
 }
